requirements.txt

babylon-python[connexion]>=29.0.0,<30.0.0

gevent>=1.3.6,<2.0.0
